fixup to compile on my Linux system. It works... kind of. Emulation seems stuck at 150% normal speed for some reason. Still, emulation works.

+++ src/sdl/debugger.cpp
@@ -19,6 +19,7 @@
 #include <stdio.h>
 #include <stdlib.h>
 #include <string.h>
+#include <stdint.h>
 
 #include "GBA.h"
 #include "Port.h"
@@ -950,9 +951,9 @@ void debuggerBreakOnWrite(u32 *mem, u32 oldvalue, u32 value, int size)
 {
   u32 address = 0;
   if(mem >= (u32*)&workRAM[0] && mem <= (u32*)&workRAM[0x3ffff])
-    address = 0x2000000 + ((u32)mem - (u32)&workRAM[0]);
+    address = 0x2000000 + ((uintptr_t)mem - (uintptr_t)&workRAM[0]);
   else
-    address = 0x3000000 + ((u32)mem - (u32)&internalRAM[0]);
+    address = 0x3000000 + ((uintptr_t)mem - (uintptr_t)&internalRAM[0]);
 
   if(size == 2)
     printf("Breakpoint (on write) address %08x old:%08x new:%08x\n",
